+++ airflow/task/task_runner/__init__.py
@@ -37,7 +37,7 @@ def get_task_runner(local_task_job):
     if _TASK_RUNNER == "StandardTaskRunner":
         return StandardTaskRunner(local_task_job)
     elif _TASK_RUNNER == "CgroupTaskRunner":
-        from airflow.contrib.task_runner.cgroup_task_runner import CgroupTaskRunner
+        from airflow.task.task_runner.cgroup_task_runner import CgroupTaskRunner
         return CgroupTaskRunner(local_task_job)
     else:
         raise AirflowException("Unknown task runner type {}".format(_TASK_RUNNER))
